BikelaneCategories: Rename file
